pyproject.toml
[project]
name = "spark-preprocessor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "camhpj", email = "camhpjohnson@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "sqlglot>=27.28.0,<27.29.dev0",
    "sqlmesh==0.228.1",
    "structlog>=25.5.0",
    "ydata-profiling>=4.18.0",
]

[project.scripts]
spark-preprocessor = "spark_preprocessor:main"

[project.optional-dependencies]
duckdb = ["duckdb>=1.0.0", "pyarrow>=15.0.0"]
databricks = ["databricks-sql-connector>=3.0.0", "pyspark>=3.5.0"]

[build-system]
requires = ["uv_build>=0.8.24,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "commitizen>=4.11.0",
    "duckdb>=1.0.0",
    "faker>=25.0.0",
    "ipython>=9.9.0",
    "jupyterlab>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "ty>=0.0.11",
]

[tool.ruff]
target-version = "py313"
exclude = ["tests", "*.ipynb", "**/*.ipynb"]

[tool.ty.src]
include = ["src"]
exclude = ["tests", "**/*.ipynb"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=spark_preprocessor",
    "--cov-report=term-missing:skip-covered",
    "--cov-fail-under=85",
]

[tool.coverage.run]
branch = true
source = ["spark_preprocessor"]

[tool.coverage.report]
skip_empty = true
omit = ["tests/*"]
